+++ include/JSystem/JKernel/JKRArchive.h
@@ -171,7 +171,7 @@ class JKRArchive : public JKRFileLoader {
     virtual void* fetchResource(void* resourceBuffer, u32 bufferSize, SDIFileEntry* entry, u32* resSize,
                                 JKRExpandSwitch expandSwitch) = 0; // _44
 
-    JKRArchive(s32, EMountMode);
+    JKRArchive(intptr_t, EMountMode);
     JKRArchive();
     JKRArchive(const char* p1, EMountMode mountMode);
     ~JKRArchive();
@@ -193,8 +193,8 @@ class JKRArchive : public JKRFileLoader {
     static JKRArchive* mount(void*, JKRHeap*, EMountDirection);
     static JKRArchive* mount(s32, EMountMode, JKRHeap*, EMountDirection);
     static void* getGlbResource(u32 type, const char* name, JKRArchive* archive);
-    static JKRArchive* check_mount_already(s32);
-    static JKRArchive* check_mount_already(s32, JKRHeap*);
+    static JKRArchive* check_mount_already(intptr_t);
+    static JKRArchive* check_mount_already(intptr_t, JKRHeap*);
 
     SDIDirEntry* findResType(u32) const;
     SDIFileEntry* findTypeResource(u32, u32) const;
@@ -234,7 +234,7 @@ class JKRArchive : public JKRFileLoader {
     // _00-_38 = JKRFileLoader
     JKRHeap* mHeap;                  // _38
     u8 mMountMode;                   // _3C
-    s32 mEntryNum;                   // _40
+    intptr_t mEntryNum;              // _40
     SArcDataInfo* mArcInfoBlock;     // _44
     SDIDirEntry* mDirectories;       // _48
     SDIFileEntry* mFileEntries;      // _4C
@@ -292,7 +292,7 @@ struct JKRCompArchive : public JKRArchive {
 
     // _00     = VTBL
     // _00-_5C = JKRArchive
-    u32 _60;                 // _60
+    uintptr_t _60;           // _60
     JKRAramBlock* mAramPart; // _64
     u32 _68;                 // _68
     JKRFile* mDvdFile;       // _6C
